feat: move interfaces to model directory Signed-off-by: Philippe Martin <phmartin@redhat.com>
